+++ packages/components/src/range-control/types.ts
@@ -202,7 +202,8 @@ export type RangeControlProps = Pick<
 		/**
 		 * Start opting into the larger default height that will become the default size in a future version.
 		 *
-		 * @default false
+		 * @deprecated Default behavior since WordPress 7.1. Prop can be safely removed.
+		 * @ignore
 		 */
 		__next40pxDefaultSize?: boolean;
 		/**
@@ -233,6 +234,7 @@ export type RangeControlProps = Pick<
 		 * Do not throw a warning for the deprecated 36px default size.
 		 * For internal components of other components that already throw the warning.
 		 *
+		 * @deprecated Default behavior since WordPress 7.1. Prop can be safely removed.
 		 * @ignore
 		 */
 		__shouldNotWarnDeprecated36pxSize?: boolean;